+++ include/reactphysics3d/collision/HalfEdgeStructure.h
@@ -122,6 +122,9 @@ class HalfEdgeStructure {
         /// Return a given vertex
         const Vertex& getVertex(uint32 index) const;
 
+        /// Reserve some memory for vertices, faces and edges
+        void reserve(uint32 facesCapacity, uint32 verticesCapacity, uint32 edgesCapacity);
+
         /// Return a string representation of the half-edge structure
         std::string to_string() const;